added pydantic requirement
